designated data filtering file instead of user-chosen
If the whole data folder and the subject number were given, the specific EEG data file should be analyzed right away instead of asking the user to input it again. MATLAB performs misbehaviors when "+" is incorporated in strings, resulting in ".set" disappearing when the file name was passed into the pop_loadset() function. One solution is passing the file name into the char() function.